Checkboxes re-run Rules Engine (#295)
* created Object to hold usageData, refactored Inputs function to eliminate redundancies


* extracted rules-engine code to app/utils/rules-engine.ts


* extracted data parsing functions to /app/utils/data-parser.ts


* comment cleanup


* extracted file upload functionality


* need to figure out how to trigger recalc

* extracted date and weather parsing logic to date-temp-util

* Pseudo code for updating billing records

* extract date and weather parsing to date-temp-util

* fix conditional display logic for EnergyUseHistory component

* add actionCounter, return pyodide Parser call to action, add comments

* create EnergyUseUpload component, rename Submit button as Re-calculate

* shell for new route for /recalculate

* on checkbox, run autonomous executeRoundtripAnalyticsFromFormJs() separate from form single.tsx's action.


* linting

* cleanup after team review, tabs->spaces

---------
