+++ heat-stack/app/components/ui/heat/CaseSummaryComponents/EnergyUseHistory.tsx
@@ -1,9 +1,10 @@
 import { Upload } from 'lucide-react'
 
 import { Button } from '#/app/components/ui/button.tsx'
-import { type UsageDataSchema } from '#/types/types.ts'; 
+import { BillingRecordsSchema, type UsageDataSchema } from '#/types/types.ts'; 
 import { AnalysisHeader } from './AnalysisHeader.tsx'
 import { EnergyUseHistoryChart } from './EnergyUseHistoryChart.tsx'
+import { EnergyUseUpload } from './EnergyUseUpload.tsx';
 
 // import { FieldMetadata, useForm } from '@conform-to/react'
 // import { Form, useLocation } from '@remix-run/react'
@@ -11,11 +12,29 @@ import { EnergyUseHistoryChart } from './EnergyUseHistoryChart.tsx'
 // import { Input } from '#/app/components/ui/input.tsx'
 // import { Label } from '#/app/components/ui/label.tsx'
 
+interface EnergyUseHistoryProps 
+{
+	lastResult: any;
+	parsedLastResult: Map<any, any> | undefined;
+	usage_data: UsageDataSchema;
+	recalculateFn: (
+		parsedLastResult: Map<any, any> | undefined,
+		billingRecords: BillingRecordsSchema,
+		parsedAndValidatedFormSchema: any,
+        convertedDatesTIWD: any,
+        state_id: any,
+        county_id: any
+	) => void;
+	show_usage_data: boolean
+}
+
 export function EnergyUseHistory({
+	lastResult,
+	parsedLastResult,
 	usage_data,
-}: {
-	usage_data: UsageDataSchema
-}) {
+	recalculateFn,
+	show_usage_data
+}: EnergyUseHistoryProps) {
 	const titleClass = 'text-5xl font-extrabold tracking-wide mt-10'
 	// const subtitleClass = 'text-2xl font-semibold text-zinc-950 mt-9'
 
@@ -23,28 +42,13 @@ export function EnergyUseHistory({
 		<div>
 			<h2 className={`${titleClass} pb-6`}>Energy Use History</h2>
 
-			<input
-				id="energy_use_upload"
-				aria-label="Upload your energy billing company's bill."
-				accept=".xml,.csv,application/xml,text/xml,text/csv,application/csv,application/x-csv,text/comma-separated-values,text/x-comma-separated-values"
-				name="energy_use_upload"
-				type="file"
-			/>
-			<Button type="submit">
-				<Upload className="mr-2 h-4 w-4" /> Upload
-			</Button>
+			
 
-			<a
-				className="inline-flex items-center gap-1 rounded-md border border-transparent bg-secondary px-2.5 py-0.5 text-xs font-semibold text-secondary-foreground transition-colors hover:bg-secondary/80 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2"
-				href="https://github.com/codeforboston/home-energy-analysis-tool/issues/162#issuecomment-2246594484"
-			>
-				Get example file here
-			</a>
-
-			{usage_data && (
+			{show_usage_data && (
 				<>
 					<AnalysisHeader usage_data={usage_data} />
-					<EnergyUseHistoryChart usage_data={usage_data} />
+					<EnergyUseHistoryChart usage_data={usage_data} lastResult={lastResult} parsedLastResult={parsedLastResult} recalculateFn={recalculateFn}/>
+					<Button type="submit">Re-calculate</Button>
 				</>
 			)}
 		</div>
